Move definitions
